feat: move db config file to infrastructure root folder
